% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\docType{data}
\name{data_bc_demo}
\alias{data_bc_demo}
\title{Breast Cancer Demograhic Information}
\format{data_bc_demo: demographic information on the patients, obtained by Haibe-Kains et al. (2012)}
\usage{
data_bc_demo
}
\description{
Breast Cancer Demograhic Information
}
\details{
Demographic information obtained by the curated breast cancer data (Haibe-Kains et al., 2012).
}
\keyword{datasets}
