% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFA_R.R
\docType{data}
\name{data_immune}
\alias{data_immune}
\title{Immune System Data}
\format{
A list of two matrices, each with 63 columns and 285 and 140 rows, respectively.
}
\usage{
data_immune
}
\description{
A data set used by De Vito et al. (2019) as an illustrative example.
}
\examples{
\dontrun{
The commands below show how the dataset was obtained from libraries on the Bioconductor repository.
source("http://bioconductor.org/biocLite.R")
biocLite(c("limma", "curatedOvarianData", "RTCGAToolbox"), suppressUpdates=TRUE)
library(curatedOvarianData)
library(RTCGAToolbox)
data(package="curatedOvarianData")
data(GSE20565_eset)
data(GSE9891_eset)
im_response <- c(
 "TP53",
 "TUBA3C","PRKACG","FGF6","FGF23","FGF22","FGF20","ASB4","TUBB1","LAT","ULBP1","NCR1",
 "SIGLEC5","CD160","KLRD1","NCR3","TRIM9","FGF18","ICOSLG",
 "MYH2","C9","MBL2",
 "GRIN2B","POLR2F","CSF2","IL5","CRP","C8A","SPTA1","GRIN2A","CCR6","FGA","LBP",
 "DUSP9","FCN2","PRKCG","ADCY8","IL5RA","GRIN1","C8B",
 "GH2","TNFSF18","GRIN2D","FGB","PRL","SPTBN5","CD70","FGG","RASGRF1","IFNG",
 "SPTBN4","TRIM10","ACTN2","LTA","TNFSF11","GRIN2C","CAMK2B","SPTB","IL1A","TNFRSF13B",
 "ITGA2B","CAMK2A","TRIM31","EREG")
GSE20_eset <- t(as.matrix(GSE20565_eset))
GSE98_eset <- t(as.matrix(GSE9891_eset))
leD <- length(im_response)
mat1 <- matrix(sapply(1:leD, function(i) which(colnames(GSE20_eset)==im_response[i])))
mat2 <- matrix(sapply(1:leD, function(i) which(colnames(GSE98_eset)==im_response[i])))
GSE20 <- matrix(c(GSE20_eset[,mat1]), 140, leD)
GSE98 <-matrix(c(GSE98_eset[,mat2]), 285, leD)
data_immune <- list(GSE98, GSE20)}
}
\keyword{datasets}
