% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\docType{data}
\name{data_subtype}
\alias{data_subtype}
\title{Breast Cancer Subtypes}
\format{
X_s_demo: the demographic information of the patients (Haibe-Kains et al., 2012) after removing the normal-like subtype

f_scoretot_sub: the estimated common factor score, as defined in De Vito et al. (2020), after removing
the normal-like subtype
}
\usage{
data_subtype
}
\description{
The resulting analysis obtained to perform the breast cancer molecular subtypes, as described in De Vito et al. (2020) in the case study application.
}
\keyword{datasets}
