% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFA_R.R
\name{ecm_msfa}
\alias{ecm_msfa}
\title{Estimates the parameters of a MSFA model}
\usage{
ecm_msfa(
  X_s,
  start,
  nIt = 50000,
  tol = 10^-7,
  constraint = "block_lower2",
  robust = FALSE,
  corr = TRUE,
  mcd = FALSE,
  trace = TRUE
)
}
\arguments{
\item{X_s}{List of lenght \eqn{S}{S}, corresponding to number of different studies considered.
Each element of the list contains a data matrix, with the same number of columns \eqn{P}{P} for all the studies.}

\item{start}{A list containing the slots \code{Phi}, \code{Lambda_s} and \code{Psi_s}, containing the starting
values for the matrix  \code{Phi} of common factor loadings, of size \eqn{P \times K}{P x K}, for
the matrices \code{Lambda_s} of study-specific factor loadings, a list of size \eqn{S}{S}  where each element
contains a matrix with \eqn{P \times J_s}{P x J_s}, and finally for the study-specific matrices of uniquenesses,
a list of size \eqn{S}{S}, where each element contains a vector of length \eqn{P}{P}.
Note that a suitable list of this kind is produced by \code{start_msfa}.}

\item{nIt}{Maximum number of iterations for the ECM algorithm. Default is 50000.}

\item{tol}{Tolerance for declaring convergence of the ECM algorithm. Default is 10^-7.}

\item{constraint}{Constraint for ensuring identifiability. The default is "block_lower2", which
corresponds to the main proposal of De Vito et al. (2018). An alternative identification
strategy is triggered by  "block_lower1"; this is more restrictive but may work also with smaller
number of variables. Again, the latter strategy is mentioned in De Vito et al. (2018).}

\item{robust}{If \code{TRUE}, robust covariance matrix is used in place of the sample covariance. Default
is \code{FALSE}.}

\item{corr}{If \code{TRUE}, the analysis will employ the correlation matrix instead of the covariance matrix.}

\item{mcd}{If \code{TRUE}, the robust estimator used for the covariance is the same proposed in Pison et al. (2003),
otherwise the default value of the function \code{CovRob} of the \code{robust} library is employed. Default is
\code{FALSE}.}

\item{trace}{If \code{TRUE} then trace information is being printed every 1000 iterations of the ECM algorithm.}
}
\value{
A list  containing the following components:
\item{\code{Phi},\code{Lambda_s}, \code{psi_s}}{the estimated model matrices.}
\item{loglik}{the value of the log likelihood function at the final estimates.}
\item{\code{AIC, BIC}}{model selection criteria at the estimate.}
\item{\code{npar}}{number of model parameters.}
\item{iter}{the number of ECM iterations performed.}
\item{constraint}{the identification constraint enforced.}
}
\description{
Maximum likelihood estimation of the MSFA model parameters via the ECM
algorithm.
}
\details{
There are two different constraints for achieving model identification,
as detailed in the reference,
though the function can also be run without such constraints (not recommended).
No checking is done on the starting value for the various model matrices,
since a suitable value for them  is produced by the function \code{start_msfa}.
}
\references{
De Vito, R., Bellio, R., Trippa, L. and Parmigiani, G. (2018). (2019). Multi-study Factor Analysis. Biometrics,  75, 337-346.

Pison, G., Rousseeuw, P.J., Filzmoser, P. and Croux, C. (2003). Robust factor analysis. Journal
of Multivariate Analysis, 84, 145-172.
}
