% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\docType{data}
\name{estim_bc}
\alias{estim_bc}
\title{Breast Cancer Estimation Results}
\format{
f_scoretot: the estimated common factor score for all the observations, as defined in De Vito et al. (2020)

Phi_sd: the estimated common factor loading matrix, a matrix of size \eqn{6362 \times 8}{6362 x 8}. This
matrix is obtained by the SD method, described in De Vito et al. (2020)

Lambda_CAL: the estimated study-specific loading matrix for the CAL dataset, a matrix of size
\eqn{6362 \times 6}{6362 x 6} obtained by the SD method.

Lambda_MAINZ: the estimated study-specific loading matrix for the CAL dataset, a matrix of size
\eqn{6362 \times 10}{6362 x 10} obtained by the SD method.
}
\usage{
estim_bc
}
\description{
The resulting analysis obtained after performing the Bayesian Multi-study Factor analysis,
as described in De Vito et al. (2020) in the case study application.
}
\keyword{datasets}
