% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\docType{data}
\name{sim_scenario2}
\alias{sim_scenario2}
\title{Setting for Scenario 2 of the Simulation Study}
\format{
Phi: the true common factor loading matrix

Phi_sd: the estimated common factor loading matrix using the SD method, as defined in De Vito et al. (2020)

Phi_op: the estimated common factor loading matrix using the OP method, as defined in De Vito et al. (2020)

RV_op: the RV coefficient between the estimated OP common factor loadings and the true common factor loadings over 50 datasets

RV_sd: the RV coefficient between the estimated SD common factor loadings and the true common factor loadings over 50 datasets
}
\usage{
sim_scenario2
}
\description{
Setting for Scenario 2 of the Simulation Study
}
\details{
This is a simulated scenario considered in De Vito et al. (2020).
The scenario considers seven different studies, with three common factors
and sixty genes (variables).
The sample size of each study is respectively (10, 15, 12, 14, 11, 11, 13).
}
\examples{
\dontrun{
The commands below illustrate the settings of Scenario 2 and how to access the
common factor loading matrix (true and estimated ones), and how to access the
RV coefficients.
S <- 7
p <- 60
k <- 3
j_s <- rep(1, 7)
n_s <- c(10, 15, 12, 14, 11, 11, 13)
theta <- rep(0, length = p)
Phi_sc2 <- data_scenario2$Phi
Phi_sd_sc2 <- data_scenario2$Phi_sd
Phi_op_sc2 <- data_scenario2$Phi_op
RV_sd_sc2 <- data_scenario2$RV_sd
RV_op_sc2 <- data_scenario2$RV_op}
}
\keyword{datasets}
