% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\docType{data}
\name{sim_scenario4}
\alias{sim_scenario4}
\title{Setting for Scenario 4 of the Simulation Study}
\format{
Phi: the true common factor loading matrix with elements greater or equal than 0.7 in absolute value

Phi_sd: the estimated common factor loading matrix using the SD method, as defined in De Vito et al. (2020). The matrix
        includes only the rows with at least one element  greater or equal than 0.7 in absolute value

Phi_op: the estimated common factor loading matrix using the OP method, as defined in De Vito et al. (2020), with at least one element greater or
equal than 0.7 in absolute value

RV_op: the RV coefficient between the estimated OP common factor loadings and the true common factor loadings over 50 datasets

RV_sd: the RV coefficient between the estimated SD common factor loadings and the true common factor loadings over 50 datasets
}
\usage{
sim_scenario4
}
\description{
Setting for Scenario 4 of the Simulation Study
}
\details{
This is a simulated scenario considered in De Vito et al. (2020). The scenario considers seven different studies, with eight common factors
and 6358 genes (variables). The sample size of each study is respectively (99, 143, 110, 120, 100, 105, 142). Here we only saved common factor
loading that are greater than 0.7 in absolute value
}
\examples{
\dontrun{
The commands below illustrate the settings of Scenario 2 and how to access the
common factor loading matrix (true and estimated ones), and how to access the
RV coefficients.
S <- 7
p <- 6358
k <- 8
j_s <- c(3, 2, 3, 6, 4, 2, 5)
n_s <- c(99, 143, 110, 120, 100, 105, 142)
theta <- rep(0, length = p)
Phi_sc4 <- data_scenario4$Phi
Phi_sd_sc4 <- data_scenario4$Phi_sd07
Phi_op_sc4 <- data_scenario4$Phi_op07
RV_sd_sc4 <- data_scenario4$RV_sd
RV_op_sc4 <- data_scenario4$RV_op}
}
\keyword{datasets}
