% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\name{sp_fa}
\alias{sp_fa}
\title{Posterior sampling for the sparse Bayesian infinite factor model.}
\usage{
sp_fa(dat, k, trace = TRUE, nprint = 1000, control = list(...), ...)
}
\arguments{
\item{dat}{Data matrix, of size \eqn{n \times p}{n x p}.}

\item{k}{Number of latent factor.}

\item{trace}{If \code{TRUE} then trace information is being printed every \code{nprint} iterations of the Gibbs sampling.
Default is \code{TRUE}.}

\item{nprint}{Frequency of tracing information. Default is every 1000 iterations.}

\item{control}{A list of hyperparameters for the prior distributions and for controlling the Gibbs sampling.
See \code{sp_fa_control}.}

\item{...}{Arguments to be used to form the default \code{control} argument if it is not supplied directly.}
}
\value{
A list, with components \code{Sigma} and \code{Lambda}, containing the posterior samples for the
covariance matrix of the model,
with dimension  \eqn{p \times p \times (nrun - code)/thin}{p x p x (nrun - burn)/thin}, and
of the loading matrix, with dimension  \eqn{p \times k \times (nrun - code)/thin}{p x k x (nrun - burn)/thin},
respectively.
}
\description{
Ported to R from Matlab code supporting the paper by Bhattacharya and Dunson (2011).
Courtesy of A. Bhattacharya.
}
\references{
Bhattacharya, A. and Dunson, D.B. (2011). Sparse Bayesian infinite factor models. Biometrika,
98, p. 291-306.
}
