% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\name{sp_fa_control}
\alias{sp_fa_control}
\title{Auxiliary function for \code{sp_fa}.}
\usage{
sp_fa_control(
  nrun = 30000,
  burn = 20000,
  thin = 1,
  nu = 3,
  asigma = 1,
  bsigma = 0.3,
  a1 = 2.1,
  b1 = 1,
  a2 = 2.1,
  b2 = 1
)
}
\arguments{
\item{nrun}{Number of posterior simulations. Default is 30000.}

\item{burn}{Burn-in trials. Default is 20000.}

\item{thin}{Thinning of posterior samples. Default is 1 (no thinning).}

\item{nu}{Parameter entering the gamma distribution assumed for phi. Default is 3.}

\item{asigma}{Shape parameter for the gamma distribution assumed for 1/sigma^2. Default is 1.}

\item{bsigma}{Scale parameter for the gamma distribution assumed for 1/sigma^2. Default is 0.3.}

\item{a1}{Shape parameter for the gamma distribution assumed for delta1. Default is 2.1.}

\item{b1}{Scale parameter for the gamma distribution assumed for delta1. Default is 1.}

\item{a2}{Shape parameter for the gamma distribution assumed for deltal. Default is 2.1.}

\item{b2}{Scale parameter for the gamma distribution assumed for deltal. Default is 1.}
}
\description{
Set parameters for posterior sampling and prior hyperparameters for the sparse Bayesian infinite factor model. For the latter, the notation
follows closely the paper by Bhattacharya and Dunson (2011).
}
\references{
Bhattacharya, A. and Dunson, D.B. (2011). Sparse Bayesian infinite factor models. Biometrika,
98, p. 291-306.
}
