% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\name{sp_msfa}
\alias{sp_msfa}
\title{Posterior sampling for the  Bayesian MSFA model for sparse setting.}
\usage{
sp_msfa(
  X_s,
  k,
  j_s,
  trace = TRUE,
  nprint = 1000,
  outputlevel = 1,
  control = list(...),
  ...
)
}
\arguments{
\item{X_s}{List of lenght \eqn{S}{S}, corresponding to number of different studies considered.
Each element of the list contains a data matrix, with the same number of columns \eqn{P}{P} for all the studies.
Standardization is carried out by the function.}

\item{k}{Number of common factors.}

\item{j_s}{Number of study-specific factors. A vector of positive integers of length \eqn{S}{S}.}

\item{trace}{If \code{TRUE} then trace information is being printed every \code{nprint} iterations of the Gibbs sampling.
Default is \code{TRUE}.}

\item{nprint}{Frequency of tracing information. Default is every 1000 iterations.}

\item{outputlevel}{Detailed level of output data. See Details. Default is 1.}

\item{control}{A list of hyperparameters for the prior distributions and for controlling the Gibbs sampling.
See \code{sp_msfa_control}.}

\item{...}{Arguments to be used to form the default \code{control} argument if it is not supplied directly.}
}
\value{
A list  containing the  posterior samples for the model parameters.  If \code{outputlevel = 1}, the components of the list are:
\item{\code{Phi}}{Common factor loadings. An array of dimension \eqn{p \times k \times (nrun - burn)/thin}{p x k x (nrun - burn)/thin}.}
\item{\code{Lambda}}{Study-specific factor loadings. A list of arrays of dimension \eqn{p \times j_s[s] \times (nrun - code)/thin}{p x j_s[s] x (nrun - burn)/thin}.}
\item{\code{psi}}{Study-specific uniquenesses. A list of arrays of dimension  \eqn{p \times 1 \times (nrun - burn)/thin}{p x 1 x (nrun - burn)/thin}.}
\item{\code{f_s}}{Study-specific latent factors associated to common factor loadings.
A list of arrays of dimension \eqn{nrow(X_s[[s]]) \times k \times (nrun - burn)/thin}{nrow(X_s[[s]]) x k x (nrun - burn)/thin}.}
\item{\code{l_s}}{Study-specific latent factors associated to study-specific factor loadings. A list of arrays of dimension
\eqn{nrow(X_s[[s]]) \times j_s[s] \times (nrun - burn)/thin}{nrow(X_s[[s]]) x j_s[s] x (nrun - burn)/thin}.}
When instead \code{outputlevel > 1}, the arrays are replaced by posterior means. The matrices \code{SigmaPhi} or the list of
matrices \code{SigmaLambda}, containing the posterior means of the these quantities, will be returned when \code{outputlevel}
is different from 1.
}
\description{
The function implements the Gibbs sampler described in De Vito et al. (2020). The code
is suitable for small to moderate-size data, and therefore can readily reproduce some of the
results of the paper, but not those for large data which would require larger computational times.
The \code{outputlevel} argument has an important role for practical usage. A value
\code{outputlevel = 1} (the default) will save all the MCMC chains, and this would create
a rather bulky output. The option \code{outputlevel = 2} will save only the chains for the loading
matrix of common factors, whereas  option \code{outputlevel = 3} will not save any chain, reporting
in the output also the posterior means of the crossproduct of the loading matrices.
}
\references{
De Vito, R., Bellio, R., Trippa, L. and Parmigiani, G. (2020).
Bayesian Multi-study Factor Analysis for High-throughput Biological Data.
Submitted manuscript.
}
