% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\name{sp_msfa_control}
\alias{sp_msfa_control}
\title{Auxiliary function for \code{sp_msfa}.}
\usage{
sp_msfa_control(
  nrun = 30000,
  burn = 20000,
  thin = 1,
  nu = 3,
  nus = 3,
  a1 = 1.1,
  b1 = 1,
  a2 = 1.1,
  b2 = 1,
  a1s = 1.1,
  b1s = 1,
  a2s = 2.1,
  b2s = 1,
  apsi = 1,
  bpsi = 0.3
)
}
\arguments{
\item{nrun}{Number of posterior simulations. Default is 30000.}

\item{burn}{Burn-in trials. Default is 20000.}

\item{thin}{Thinning of posterior samples. Default is 1 (no thinning).}

\item{nu}{Parameter entering the gamma distribution assumed for omega. Default is 3.}

\item{nus}{Parameter entering the gamma distribution assumed for omegas. Default is 3.}

\item{a1}{Shape parameter for the gamma distribution assumed for delta1. Default is 1.1.}

\item{b1}{Scale parameter for the gamma distribution assumed for delta1. Default is 1.}

\item{a2}{Shape parameter for the gamma distribution assumed for deltal. Default is 1.1.}

\item{b2}{Scale parameter for the gamma distribution assumed for deltal. Default is 1.}

\item{a1s}{Shape parameter for the gamma distribution assumed for delta1_s. Default is 1.1.}

\item{b1s}{Scale parameter for the gamma distribution assumed for delta1_s. Default is 1.}

\item{a2s}{Shape parameter for the gamma distribution assumed for deltal_s. Default is 2.1.}

\item{b2s}{Scale parameter for the gamma distribution assumed for deltal_s. Default is 1.}

\item{apsi}{Shape parameter for the gamma distribution assumed for 1/psi. Default is 1.}

\item{bpsi}{Scale parameter for the gamma distribution assumed for 1/psi. Default is 0.3.}
}
\description{
Set parameters for posterior sampling and prior hyperparameters for the  Bayesian MSFA model
for sparse setting. The notation follows closely the paper by De Vito et al. (2020).
}
\references{
De Vito, R., Bellio, R., Trippa, L. and Parmigiani, G. (2020).
Bayesian Multi-study Factor Analysis for High-throughput Biological Data.
Submitted manuscript.
}
