% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\name{sp_OP}
\alias{sp_OP}
\title{OP algorithm for loading matrix estimation}
\usage{
sp_OP(
  Phi_sample,
  tol = 10^-7,
  itermax = 50,
  init = NULL,
  updatesample = FALSE,
  trace = TRUE
)
}
\arguments{
\item{Phi_sample}{Posterior samples for a loading matrix, a 3-dimensional array of
dimension \eqn{p \times k \times nchain}{p x k x nchain}.}

\item{tol}{Tolerance  for the OP algorithm. Default is 0.001.}

\item{itermax}{Maximum number of iterations. Default is 10.}

\item{init}{Starting point for the optimization.}

\item{updatesample}{Should the posterior samples be updated? Default is \code{FALSE}.}

\item{trace}{Print tracing information. Default is \code{TRUE}.}
}
\value{
A list containing the estimated loading matrix, plus the number of iterations
performed and the Frobenius norm of the difference between the last two iterations
of the algorithm.
}
\description{
Iterative algorithm for estimating a loading matrix from the posterior samples, resolving
the orthogonality indeterminacy.
}
