% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFA_R.R
\name{start_msfa}
\alias{start_msfa}
\title{Provides some starting values for the parameters of a MSFA model}
\usage{
start_msfa(
  X_s,
  k,
  j_s,
  constraint = "block_lower2",
  method = "adhoc",
  robust = FALSE,
  corr = FALSE,
  mcd = FALSE
)
}
\arguments{
\item{X_s}{List of lenght \eqn{S}{S}, corresponding to number of different studies considered.
Each element of the list contains a data matrix, with the same number of columns \eqn{P}{P} for all the studies.
No standardization is carried out by the function.}

\item{k}{Number of common factors.}

\item{j_s}{Number of study-specific factors. A vector of positive integers of length \eqn{S}{S}.}

\item{constraint}{Constraint for ensuring identifiability. The default is "block_lower2", which
corresponds to the main proposal of De Vito et al. (2018). An alternative identification
strategy is triggered by  "block_lower1"; this is more restrictive but may work also with smaller
number of variables.}

\item{method}{Which method should be used to find the starting values? The two possibilities are \code{"adhoc"} for
the method described in De Vito et al. (2016), and \code{"fa"} for averaging over separate study-specific FA models.
Default is \code{"adhoc"}.}

\item{robust}{If \code{TRUE}, robust covariance matrix is used in place of the sample covariance. Default
is \code{FALSE}.}

\item{corr}{If \code{TRUE}, the analysis will employ the correlation matrix instead of the covariance matrix.}

\item{mcd}{If \code{TRUE}, the robust estimator used for the covariance is the same proposed in Pison et al. (2003),
otherwise the default value of the function \code{CovRob} of the \code{robust} library is employed. Default is
\code{FALSE}.}
}
\value{
A list  containing  \code{Phi},\code{Lambda_s} and  \code{psi_s}, starting values for the model matrices.
}
\description{
This is a supporting function for \code{ecm_msfa}. The method employed is documented in the reference.
}
\details{
The upper-triangular zero constraint is adopted to achieve identification,
as detailed in the reference, though the function can also be run without such constraint.
}
\references{
De Vito, R., Bellio, R., Parmigiani, G. and Trippa, L. (2019). Multi-study Factor Analysis. Biometrics,  75, 337-346.
}
