% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSFA.R
\name{ecm_msfa}
\alias{ecm_msfa}
\title{Estimates the parameters of a MSFA model}
\usage{
ecm_msfa(X_s, start = NULL, nIt = 50000, tol = 10^-10,
  block_lower = TRUE, robust = FALSE, corr = TRUE, mcd = FALSE,
  trace = TRUE)
}
\arguments{
\item{X_s}{List of lenght \eqn{S}{S}, corresponding to number of different studies considered.
Each element of the list contains a data matrix, with the same number of columns \eqn{P}{P} for all the studies.}

\item{start}{A list containing the slots \code{Phi}, \code{Lambda_s} and \code{Psi_s}, containing the starting
values for the matrix  \code{Phi} of common factor loadings, of size \eqn{P \times K}{P x K}, for
the matrices \code{Lambda_s} of study-specific factor loadings, a list of size \eqn{S}{S}  where each element
contains a matrix with \eqn{P \times J_s}{P x J_s}, and finally for the study-specific matrices of uniquenesses,
a list of size \eqn{S}{S}, where each element contains a diagonal matrix with \eqn{P \times P}{P x P}.
Note that a suitable list of this kind is produced by \code{start_msfa}.}

\item{nIt}{Maximum number of iterations for the ECM algorithm. Default is 20000.}

\item{tol}{Tolerance for declaring convergence of the ECM algorithm. Default is 10^-7.}

\item{block_lower}{Should the upper-triangular zero constraint be enforced? Default is \code{TRUE}
(strongly suggested).}

\item{robust}{If \code{TRUE}, robust covariance matrix is used in place of the sample covariance. Default
is \code{FALSE}.}

\item{corr}{If \code{TRUE}, the analysis will employ the correlation matrix instead of the covariance matrix.}

\item{mcd}{If \code{TRUE}, the robust estimator used for the covariance is the same proposed in Pison et al. (2003),
otherwise the default value of the function \code{CovRob} of the \code{robust} library is employed. Default is
\code{FALSE}.}

\item{trace}{If \code{TRUE} then trace information is being printed every 1000 iterations of the ECM algorithm.}
}
\value{
A list  containing the following components:
\item{\code{Phi},\code{Lambda_s}, \code{Psi_s}}{the estimated model matrices.}
\item{loglik}{the value of the log likelihood function at the final estimates.}
\item{\code{AIC, BIC}}{model selection criteria at the estimate.}
\item{\code{npar}}{number of model parameters.}
\item{iter}{the number of ECM iterations performed.}
}
\description{
Maximum likelihood estimation of the MSFA model parameters via the ECM
algorithm.
}
\details{
The upper-triangular zero constraint is adopted to achieve identification,
as detailed in the reference, though the function can also be run without such constraint.
No checking is done on the starting value for the various model matrices,
since a suitable value for them  is produced by the function \code{start_msfa}.
}
\references{
De Vito, R., Bellio, R., Trippa, L. and Parmigiani, G. (2016). Multi-study Factor Analysis,
arXiv:1611.06350. \url{https://arxiv.org/abs/1611.06350}

Pison, G., Rousseeuw, P.J., Filzmoser, P. and Croux, C. (2003). Robust factor analysis. Journal
of Multivariate Analysis, 84, 145-172.
}
