% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.R
\name{sp_OP}
\alias{sp_OP}
\title{OP algorithm for loading matrix estimation}
\usage{
sp_OP(Phi, tol = 10^-3, itermax = 10)
}
\arguments{
\item{Phi}{Posterior samples for a loading matrix, a 3-dimensional array of
dimension \eqn{p \times k \times nchain}{p x k x nchain}.}

\item{tol}{Tolerance  for the OP algorithm. Default is 0.001.}

\item{itermax}{Maximum number of iterations. Default is 10.}
}
\value{
A list containing the estimated loading matrix, plus the number of iterations
performed and the Frobenius norm of the difference between the last two iterations
of the algorithm.
}
\description{
Iterative algorithm for estimating a loading matrix from the posterior samples, resolving
the orthogonality indeterminacy.
}
